Consensus: MOVEONLY: Move to blockverify.cpp:
from main.cpp:
-CheckBlockHeader
-ContextualCheckBlockHeader
-IsSuperMajority

from pow.cpp:
-CalculateNextWorkRequired
-CheckProofOfWork
-GetNextWorkRequired
